innovation-lab-examples/launch-your-a2a-research-team/requirements.txt
uagents-core[a2a]==0.4.5a2
a2a-sdk[http-server]>=0.3.20,<0.4.0
python-dotenv
google-genai
uvicorn
